+++ test/reducers/modal.spec.js
@@ -38,6 +38,10 @@ describe('reducers/modal', () => {
     const state = {};
     deepFreeze(state);
 
-    expect(modal(state, {})).to.be.deep.eql(state);
+    expect(modal(state, {})).to.be.deep.eql({
+      title: null,
+      body: null,
+      open: false,
+    });
   });
 });
